requirements.txt
python-telegram-bot==20.8
stripe==9.7.0
requests==2.31.0